finish splitting up file
